Fixed deleting files instead of just modded dlls

+++ PizzaOven/ModLoader.cs
@@ -26,7 +26,7 @@ namespace PizzaOven
             var dlls = new List<string>(new string[] { "fmod.dll", "fmod-gamemaker.dll", "fmodstudio.dll", "gameframe_x64.dll", "steam_api.dll",
             "steam_api64.dll", "steamworks_x64.dll"});
             foreach (var file in Directory.GetFiles($"{Global.config.ModsFolder}", "*", SearchOption.TopDirectoryOnly))
-                if (!dlls.Contains(Path.GetFileName(file).ToLowerInvariant()))
+                if (Path.GetExtension(file).ToLowerInvariant() == ".dll" && !dlls.Contains(Path.GetFileName(file).ToLowerInvariant()))
                     File.Delete(file);
             foreach (var directory in Directory.GetDirectories($"{Global.config.ModsFolder}{Global.s}sound{Global.s}Desktop"))
                 if (Directory.GetFiles(directory).Length == 0 && Directory.GetDirectories(directory).Length == 0)
